+++ radixdbmatch.c
@@ -27,7 +27,7 @@ int main(int argc, char **argv) {
   struct radixdb db;
   int fd;
   const char *val, *key;
-  uint32_t vlen, klen;
+  size_t vlen, klen;
 #ifdef _WIN32
   HANDLE hFile, hMapping;
 #endif
